Move data types into a dedicated submod data
